Remote logs_proxy from collector
This endpoint proxied OTLP/grpc logs requests directly to ingestor.
Instead, users should use the logs_transfer mechanism that writes to the
local store which is transferred to ingestor which is more efficient.
