+++ collector/otlp/logs_transfer_test.go
@@ -88,3 +88,80 @@ func TestLogsService_Overloaded(t *testing.T) {
 	keys := store.PrefixesByAge()
 	require.Equal(t, 0, len(keys))
 }
+
+type fakeHealthChecker struct {
+	healthy bool
+}
+
+func (f fakeHealthChecker) IsHealthy() bool { return f.healthy }
+
+var rawlog = []byte(`{
+	"resourceLogs": [
+		{
+			"resource": {
+				"attributes": [
+					{
+						"key": "source",
+						"value": {
+							"stringValue": "hostname"
+						}
+					}
+				],
+				"droppedAttributesCount": 1
+			},
+			"scopeLogs": [
+				{
+					"scope": {
+						"name": "name",
+						"version": "version",
+						"droppedAttributesCount": 1
+					},
+					"logRecords": [
+						{
+							"timeUnixNano": "1669112524001",
+							"observedTimeUnixNano": "1669112524001",
+							"severityNumber": 17,
+							"severityText": "Error",
+							"body": {
+								"kvlistValue": {
+									"values": [
+									  {
+										"key": "message",
+										"value": {
+										  "stringValue": "something worth logging"
+										}
+									  },
+									  {
+										"key": "utf8message",
+										"value": {
+										  "stringValue": "🔥 parse please"
+										}
+									  },
+									  {
+										"key": "kusto.table",
+										"value": {
+										  "stringValue": "ATable"
+										}
+									  },
+									  {
+										"key": "kusto.database",
+										"value": {
+										  "stringValue": "ADatabase"
+										}
+									  }
+									]
+								  }
+							},
+							"droppedAttributesCount": 1,
+							"flags": 1,
+							"traceId": "",
+							"spanId": ""
+						}
+					],
+					"schemaUrl": "scope_schema"
+				}
+			],
+			"schemaUrl": "resource_schema"
+		}
+	]
+}`)
